% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppc.R
\name{plot_ppc}
\alias{plot_ppc}
\title{Plot Posterior Predictive Check with Bayesian p-value}
\usage{
plot_ppc(fit, obs_var = "T_obs_sum", sim_var = "T_sim_sum")
}
\arguments{
\item{fit}{A CmdStanR fit object.}

\item{obs_var}{Name of the observed test statistic variable (string).}

\item{sim_var}{Name of the simulated test statistic variable (string).}
}
\value{
A ggplot2 object showing the PPC scatter plot.
}
\description{
Plot Posterior Predictive Check with Bayesian p-value
}
