% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traceplot.R
\name{plot_traceplot}
\alias{plot_traceplot}
\title{Plot Traceplots for MCMC Draws Array}
\usage{
plot_traceplot(draws, parameters = NULL, ncol = 3)
}
\arguments{
\item{draws}{A 3D array of MCMC samples (iterations x chains x parameters).}

\item{parameters}{Optional vector of parameter names to include in the plot.}

\item{ncol}{Number of columns in facet layout (default is 3).}
}
\value{
A ggplot2 object showing traceplots.
}
\description{
Plot Traceplots for MCMC Draws Array
}
