% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traceplot2.R
\name{plot_traceplot2}
\alias{plot_traceplot2}
\title{Plot traceplots for MCMC draws array}
\usage{
plot_traceplot2(draws, mcmc_obj, parameters = NULL, ncol = 3)
}
\arguments{
\item{draws}{A 3D array of MCMC samples (iterations x chains x parameters).}

\item{mcmc_obj}{a cmdstanr MCMC object}

\item{parameters}{Optional vector of parameter names to include in the plot.}

\item{ncol}{Number of columns in facet layout (default is 3).}
}
\value{
A ggplot2 object showing traceplots with ESS and R-hat.
}
\description{
Plot traceplots for MCMC draws array
}
